requirements.txt
Flask==2.2.5
gunicorn==20.1.0
tensorflow-cpu==2.9.3
joblib==1.2.0
numpy==1.21.6
h5py==3.7.0
scikit-learn==1.1.3
